% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_coast_low_res.R
\docType{data}
\name{BC_coast_low_res}
\alias{BC_coast_low_res}
\title{BC Coast Low Res}
\format{
An object of class \code{sf} (simple feature) with 783 features and 8 fields.
The projected CRS is NAD83 / BC Albers (EPSG:3005).
}
\source{
Obtained from the Canadian Hydrographic Service.
}
\usage{
BC_coast_low_res
}
\description{
A low resolution sf polygon shape file for the BC coastline.
This is an appropriate resolution for coastwide maps but does not have sufficient resolution for maps of small areas.
}
\keyword{datasets}
