% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_split_violin.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{geom_split_violin}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "width",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()} or \code{aes_()}. If specified and \code{inherit.aes = TRUE} (the default), it is combined with the default mapping at the top level of the plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options: If NULL, the default, the data is inherited from the plot data as specified in the call to \code{ggplot()}. If a \code{data.frame}, that will be used as the layer data. A \code{data.frame} (or other object) will override the plot data. All objects will be fortified to produce a data frame. See \code{fortify()} for which variables will be created. A function will be called with a single argument, the plot data. The return value must be a \code{data.frame}, and will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{draw_quantiles}{A numeric vector of values between 0 and 1 to draw quantile lines at.}

\item{trim}{If \code{FALSE} (the default), all data are included in the violins. If \code{TRUE}, data are trimmed to the range of the plot.}

\item{scale}{If \code{"area"} (the default), all violins have the same area (before trimming). If \code{"count"}, areas are scaled proportionally to the number of observations. If \code{"width"}, all violins have the same maximum width.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{Logical. Should this layer be included in the legends? \code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. \code{borders()}.}
}
\value{
A ggplot layer.
}
\description{
Create a split violin plot.
}
\seealso{
\code{\link[ggplot2]{layer}}, \code{\link[ggplot2]{geom_violin}}
}
