% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_split_violin.R
\docType{data}
\name{GeomSplitViolin}
\alias{GeomSplitViolin}
\title{GeomSplitViolin}
\format{
An object of class \code{GeomSplitViolin} (inherits from \code{GeomViolin}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomSplitViolin
}
\arguments{
\item{self}{A ggproto object.}

\item{data}{A data frame.}

\item{draw_quantiles}{A numeric vector of values between 0 and 1 to draw quantile lines at.}

\item{...}{Other arguments passed on to methods.}
}
\value{
A ggplot grob.
}
\description{
A custom ggproto object for creating split violin plots.
}
\seealso{
\code{\link[ggplot2]{GeomViolin}}, \code{\link[ggplot2]{GeomPolygon}}
}
\keyword{datasets}
