% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale data function}
\usage{
unscale(x_scaled, x)
}
\arguments{
\item{x_scaled}{Vector, data on the standardized scale}

\item{x}{Vector, the original data that was standardized}
}
\value{
A vector of data in the unstandardized scale
}
\description{
This function takes data on a standardized (mean = 0, sd = 1) scale and converts it back to the original scale.
This is useful for converting predictions from models that used standardized variables back to their original scale.
}
