% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_theme.R
\name{theme_pt}
\alias{theme_pt}
\title{A custom ggplot theme}
\usage{
theme_pt()
}
\description{
This function applies a custom theme to a ggplot object. It is based on
theme_bw(), but with grid lines removed and facet boxes white with no border.
}
\examples{
# Create a basic ggplot
p <-  ggplot2::ggplot(mtcars, ggplot2::aes(mpg, hp)) + ggplot2::geom_point()

# Apply the custom theme
p + theme_pt()
}
